Starting to feed code from the editor to the app
