+++ Continuous.Client.VisualStudio/packages.config
@@ -1,5 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="Microsoft.CodeAnalysis.Analyzers" version="1.1.0" targetFramework="net46" />
+  <package id="Microsoft.CodeAnalysis.Common" version="1.1.1" targetFramework="net46" />
+  <package id="Microsoft.CodeAnalysis.EditorFeatures.Text" version="1.1.1" targetFramework="net46" />
+  <package id="Microsoft.CodeAnalysis.Workspaces.Common" version="1.1.1" targetFramework="net46" />
+  <package id="Microsoft.Composition" version="1.0.27" targetFramework="net46" />
   <package id="Microsoft.VisualStudio.Imaging" version="14.1.24720" targetFramework="net452" />
   <package id="Microsoft.VisualStudio.OLE.Interop" version="7.10.6070" targetFramework="net452" />
   <package id="Microsoft.VisualStudio.Shell.14.0" version="14.1.24720" targetFramework="net452" />
@@ -20,4 +25,21 @@
   <package id="Microsoft.VisualStudio.Validation" version="14.1.111" targetFramework="net452" />
   <package id="Microsoft.VSSDK.BuildTools" version="14.1.24720" targetFramework="net452" developmentDependency="true" />
   <package id="Newtonsoft.Json" version="7.0.1" targetFramework="net452" />
+  <package id="System.Collections" version="4.0.0" targetFramework="net46" />
+  <package id="System.Collections.Immutable" version="1.1.37" targetFramework="net46" />
+  <package id="System.Diagnostics.Debug" version="4.0.0" targetFramework="net46" />
+  <package id="System.Globalization" version="4.0.0" targetFramework="net46" />
+  <package id="System.IO" version="4.0.0" targetFramework="net46" />
+  <package id="System.Linq" version="4.0.0" targetFramework="net46" />
+  <package id="System.Reflection" version="4.0.0" targetFramework="net46" />
+  <package id="System.Reflection.Extensions" version="4.0.0" targetFramework="net46" />
+  <package id="System.Reflection.Metadata" version="1.1.0" targetFramework="net46" />
+  <package id="System.Reflection.Primitives" version="4.0.0" targetFramework="net46" />
+  <package id="System.Resources.ResourceManager" version="4.0.0" targetFramework="net46" />
+  <package id="System.Runtime" version="4.0.0" targetFramework="net46" />
+  <package id="System.Runtime.Extensions" version="4.0.0" targetFramework="net46" />
+  <package id="System.Runtime.InteropServices" version="4.0.0" targetFramework="net46" />
+  <package id="System.Text.Encoding" version="4.0.0" targetFramework="net46" />
+  <package id="System.Text.Encoding.Extensions" version="4.0.0" targetFramework="net46" />
+  <package id="System.Threading" version="4.0.0" targetFramework="net46" />
 </packages>
